+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Pixel gifs
   - Power options
   - Builtin animations
-- In-progress simulators for GA402, GU604 animatrix
+- In-progress simulators for GA402, GU604 animatrix, optional build and takes a single arg
 - Add `model_override` option to anime config, this is handy for forcing a model for "Unknown" anime, and for simulators
 ### Changed
 - Move FX506HC to FX506H in arua DB to catch full series of this range